Handle 'ptp4l' restarts
If 'ptp4l' restarts there might be insufficient data to query from the result dictionary using ptp keywords keys which causing the code to break (for example if using a system command and applying the change).

The fix is to check that the amount of items in the result dictionary is equal to the amount of ptp keywords keys used for the query.

Story: 2008529
Task: 41823

Signed-off-by: Udi Schwager <udi.schwager@windriver.com>
Change-Id: I0718afd9d0681b391e3426c9ab653eb9184710d0