Remove pod resource spec for ptp-notification
The daemonset of ptp-notification was configured with explicit resource
requests. The ptp-notification application is configured to run on
dedicated platform cores and was improperly causing them to count
against the overall k8s not resources. Removing the explicit resource
configuration will correct this. It is not required for the normal
operation of this application.

Closes-Bug: 1927971

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: If7414052c70b55947d05b5a46d32eaa003ea56eb