Remove --use-deprecated legacy-resolver flag
The pip team is planning to depreciate the
'--use-deprecated legacy-resolver'install option.
While a firm date has yet to be set, We need to be
ready for the change.

Test Plan:

PASS: Verified that tox runs without issues
PASS: Generate ptp-notification package.
PASS: Upload and apply ptp-notification package and check if pod
      is running.

Closes-bug:1966073

signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>
Change-Id: I854cfb8b72bb9af069b8c0c378f934cf189c6345