[PTP] Handle NICs with a single PHC device
Issue: The logic to determine the path for the PHC device does not work
properly when a NIC has only one PHC device.

The way to determine the OS clock accuracy/offest uses the phc_ctl
command to read the difference between the PHC and the OS clock. In
order to do this from the container, we need the path for the PHC
device.

Added logic to try the 0th interface of the NIC if there is no device
found on the initial phc_interface. Some NICs with a single PHC only
present it on the base interface.

Test plan:
PASS: Build and deploy notificationservice-base container image
PASS: Monitor OS clock offset using the non-0th interface and verify
PHC device is still found

Story: 2010056
Task: 46323
Change-Id: Idf77f9abd73e6096c832d2905b26103061104f4d