[PTP] Automatically determine list of services to monitor
Issue: Deploying ptp-notification on multiple nodes does not work
properly when each node has different ptp-instance names configured.

This change allows the ptp-notification pod to discover the ptp services
on each node that it is running on and generate a list of service names
to monitor. This also removes the requirement for users to supply an
override with a list of service names.

A given monitoring type can be disabled by supplying an override value
of "False" and ptp-notification will not track it.

Test plan:
PASS: Build and deploy ptp-notification
PASS: Automatically monitor all configured instances
PASS: Disable each instance type independently

Story: 2010056
Task: 46356

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: Ieb3ee26b9431e7ac15a7599a69ce9846dce78328