[PTP] No longer consider clockClass 7 and 135 as locked
PTP-notification was treating clockClass values 7 and 135 as indicators
of a locked status, which is incorrect as per ITU-T G.8275.1. This was
resulting in ptp-notification delivering locked status to clients when
it should instead be transitioning to holdover/freerun.

Removed clockClass 7 and 135 from the list of valid locked clockClasses.
PTP-notification is now able to properly transition states when these
classes are observed.

Test plan:
PASS: Build and deploy image
PASS: Verify correct state transitions when clockClass 6, 7, 135 are
observed

Closes-Bug: 1995011

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: Ie992603a84e25f7e4a27adc6c11b3db9b6878bed