Fix log filename where nmea is searched
When ptp-notification application is used with a ts2phc instance, the
name of nmea serial port (ttyGNSS device) is retrieved from the kernel
logs. This change fixes the look for 'dmesg' file which is not available
by replacing it by 'kern.log' where that information could be found.

Test Plan:
PASS: Updated ptp-notification-ptptracking container running in
      environment with ts2phc instance configured and running.

Closes-Bug: #1995718
Signed-off-by: Douglas Henrique Koerich <douglashenrique.koerich@windriver.com>
Change-Id: I39a147f2eb626ad843035935de334d9f3dcf33d7