Fix unhandled exception when querying ptp status
Attempting to query the status of a non-existant ptp instance or
clockClass could result in an unhandled key error exception on the
ptptracking daemon.

Added logic to handle this case on the server and report the error to
the client.

Test plan:
PASS: Build and deploy notificationservice-base and
notificationclient-base
PASS: Test v1 and v2 status pulls, subscribe, list and delete.

Closes-bug: 1997995

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I3bfe111dbd79475720ea1e710be2a23f45763ae6