Fix centos image build failure
Corrects an issue with the centos stable_docker_image files for the
notificationservice-base image which was resulting in an image build
failure.

notificationservice-base.stable_docker_image was missing the path to the
Dockerfile.

The notificationservice-base-v2.stable_docker_image file was also
incorrectly named.

Test plan:
PASS: Build centos images for notificationservice-base and
notificationservice-base-v2
PASS Build debian images for notificationservice-base and
notificationservice-base-v2

Closes-bug: 2004428

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I17984044888e4cf5e8985376e28f17e143341472