Fix logic for disabling v1 api container
The existing logic in the daemonset template was removing several of the
required volume mounts and resulting in a failure to create the pods.
The fix preserves the volume mounts even if the v1 notification is
disabled.

Test-plan:
PASS: Build and deploy ptp-notification
PASS: Enable/disable v1/v2 api containers

Story: 2010538
Task: 47557

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: Iee6dacbf36e038c61de9450ca15e5cdc135d97aa