Update ptp-notification metadata for upgrade support
This change updates the metadata of the ptp-notification application
tarball to mark the app as platform managed and eligible for automatic
upgrades. The application framework will now properly upgrade
ptp-notification when a newer version is installed on a system.

Test plan:
PASS: Build and deploy updated ptp-notification
PASS: Patch system to update ptp-notification application, verify that
app is automatically upgraded.

Story: 2010538
Task: 47583

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I3e42a58f9d01fbfdfadf2d733fc3411c6732384b