Update notificationservice-base-v1 startup script to handle upgrades
Updated the logic in ptptracking_start.py to handle a case where the
application is upgraded but helm overrides are not present.

If helm overrides are present, there is no change in behaviour.
If helm overrides are absent, attempt to locate ptp4l and phc2sys
configs in known locations and start with those.
    Print log messages indicating that the configs were auto detected
    and that helm overrides can be set if auto-detection is incorrect or
    unsuccessful.

Bonus fix: corrected the tox path used to auto detect unit tests, as
this was missed when moving from notificationservice-base to
notificationservice-base-v2

Test-plan:
PASS: Build and deploy application tarball
PASS: Pods start correctly when overrides are present and when they are
absent

Story: 2010538
Task: 47740

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I25df407effbffc03e444573233058f3d3180e706