Handle missing sysfs path for in-tree ice driver
The in-tree ice driver version 6.6.40-stx.2 does not provide the debugfs
path used for monitoring the GNSS state. This commit provides the
following changes to support this driver:

1: Update the k8s daemonset to mount the /sys/kernel/debug path instead
   of /sys/kernel/debug/ice. The "ice" directory is not present using
   the in-tree driver, so mounting the parent directory allows the
   containers to be properly started and can then check if the "ice"
   directory is present at runtime
2: Update the startup script for the v2 container to check for the
   existence of the ice debugfs directory. If it is not present, disable
   GNSS tracking in ptp-notification
3: Update the path used by GNSS monitoring to match the changed mount
   path in the daemonset so that GNSS monitoring can proceed when the
   path is present

Currently, GNSS time sources and monitoring cannot be used with the
in-tree driver. GNSS time sources require configuring the system to use
the out-of-tree ice driver.

Test plan:
Pass: Verify helm chart and container build
Pass: Verify ptp-notification deployment using the in-tree ice driver,
ensure that GNSS tracking is automatically disabled when debugfs is not
present
Pass: Verify ptp-notification deployment using the out-of-tree ice
driver, ensure that GNSS tracking is operational when debugfs is present

Story: 2011056
Task: 50940

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I73859408cf5655b58ebe2a1a78f407496c3380dd