Fix for rabbitmq address format
A package update sometime after the notificationclient 2.3.0 release has
changed the behaviour of oslo_messaging.

In rpc_helper.py get_transport(), the oslo_messaging.get_rpc_transport()
now errors out when passed an ipv4 address wrapped in "[]". This was not
previously the case.

Updated the parsing for the broker_host variable so that it is only
wrapped in brackets when it is ipv6.

Testing:
Pass: Verify operation on ipv4 systems
Pass: Verify operation on ipv6 systems

Closes-bug: 2093125

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I7211d0d4c2769547233ecc1dae0416327a19c32c