API-REF: Subscriptions do not persist on host reboot
The default storage volume that the application sidecar uses to store
subscriptions is not intended to persist on host reboots or pod
deletion.

This change adds a note to the api-ref documents indicating that it is
best practice for the consumer application to subscribe to all required
resources on application startup.

Test-plan:
N/A - Doc update only

Closes-Bug: 2111532

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I107fc0bd895acc6a8db794568c98efe76f7fd140